Made to implement Shape
